requirements.txt
opencv-python
mediapipe-silicon
numpy
mido
python-rtmidi
